+++ src/Singularity/Url/Generic.php
@@ -15,6 +15,7 @@ use DecodeLabs\Singularity\Url;
 
 class Generic implements
     Url,
+    Rebasable,
     Dumpable
 {
     use SchemeTrait;
@@ -25,6 +26,7 @@ class Generic implements
     use PathTrait;
     use QueryTrait;
     use FragmentTrait;
+    use RebaseTrait;
 
     public static function fromString(
         string $uri